+++ packages/bpk-component-chip/stories.js
@@ -16,11 +16,15 @@
  * limitations under the License.
  */
 
+/* @flow */
+
 import React from 'react';
 import { storiesOf, action } from '@storybook/react';
 
 import BpkChip from './index';
 
 storiesOf('bpk-component-chip', module).add('Default', () => (
-  <BpkChip onClose={action('Chip closing!')}>This is a chip!</BpkChip>
+  <BpkChip onClose={action('Chip closing!')} closeLabel="Close">
+    This is a chip!
+  </BpkChip>
 ));